adapt party types
